Separated transmission test module
